requirements.txt
PyQt6
Pillow
PyPDF2
pdfplumber
pytesseract
reportlab
python-docx
